+++ utils/__init__.py
@@ -1,6 +1,7 @@
 from utils.compress import compress_database, compress_folder
 from utils.dialog_buttons import DialogButtons
 from utils.dialog_icons import Icons
+from utils.excel_file import ExcelFile
 from utils.extract import extract
 from utils.file_changes import FileChanges
 from utils.ip_utils import get_server_ip_address, get_server_port, get_system_ip_address